Move support within sections, clean up update checking
The new method now iterates both the existing and updated objects at the same time (instead of separately) so we can identify moves at the same time.

Right now this handles moves by doing a delete and an insert on each location of the move. This will be done shortly, but I wanted a midpoint where *this* works before continuing on.