Don't force sort descriptors in HRC
Forcing a sort descriptor to the Core Data level, for single object requests, causes us to have to do something hacky: pick a key and pretend it's useful as a sort.

But what if that key is something that can change? We process it as a delete/insert. Which isn't good when the key we pick is something that _will_ change.

This drops the requirement and does some internal logic weirdness to keep state consistent.